// web/src/generated/models/InifiniteScrollPagnitionUserPublicInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Revornix Main Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.4
 * Contact: 1142704468@qq.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserPublicInfo } from './UserPublicInfo';
import {
    UserPublicInfoFromJSON,
    UserPublicInfoFromJSONTyped,
    UserPublicInfoToJSON,
    UserPublicInfoToJSONTyped,
} from './UserPublicInfo';

/**
 * 
 * @export
 * @interface InifiniteScrollPagnitionUserPublicInfo
 */
export interface InifiniteScrollPagnitionUserPublicInfo {
    /**
     * 
     * @type {number}
     * @memberof InifiniteScrollPagnitionUserPublicInfo
     */
    total: number;
    /**
     * 
     * @type {number}
     * @memberof InifiniteScrollPagnitionUserPublicInfo
     */
    start?: number | null;
    /**
     * 
     * @type {number}
     * @memberof InifiniteScrollPagnitionUserPublicInfo
     */
    limit: number;
    /**
     * 
     * @type {boolean}
     * @memberof InifiniteScrollPagnitionUserPublicInfo
     */
    has_more: boolean;
    /**
     * 
     * @type {Array<UserPublicInfo>}
     * @memberof InifiniteScrollPagnitionUserPublicInfo
     */
    elements: Array<UserPublicInfo>;
    /**
     * 
     * @type {number}
     * @memberof InifiniteScrollPagnitionUserPublicInfo
     */
    next_start?: number | null;
}

/**
 * Check if a given object implements the InifiniteScrollPagnitionUserPublicInfo interface.
 */
export function instanceOfInifiniteScrollPagnitionUserPublicInfo(value: object): value is InifiniteScrollPagnitionUserPublicInfo {
    if (!('total' in value) || value['total'] === undefined) return false;
    if (!('limit' in value) || value['limit'] === undefined) return false;
    if (!('has_more' in value) || value['has_more'] === undefined) return false;
    if (!('elements' in value) || value['elements'] === undefined) return false;
    return true;
}

export function InifiniteScrollPagnitionUserPublicInfoFromJSON(json: any): InifiniteScrollPagnitionUserPublicInfo {
    return InifiniteScrollPagnitionUserPublicInfoFromJSONTyped(json, false);
}

export function InifiniteScrollPagnitionUserPublicInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): InifiniteScrollPagnitionUserPublicInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'total': json['total'],
        'start': json['start'] == null ? undefined : json['start'],
        'limit': json['limit'],
        'has_more': json['has_more'],
        'elements': ((json['elements'] as Array<any>).map(UserPublicInfoFromJSON)),
        'next_start': json['next_start'] == null ? undefined : json['next_start'],
    };
}

export function InifiniteScrollPagnitionUserPublicInfoToJSON(json: any): InifiniteScrollPagnitionUserPublicInfo {
    return InifiniteScrollPagnitionUserPublicInfoToJSONTyped(json, false);
}

export function InifiniteScrollPagnitionUserPublicInfoToJSONTyped(value?: InifiniteScrollPagnitionUserPublicInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'total': value['total'],
        'start': value['start'],
        'limit': value['limit'],
        'has_more': value['has_more'],
        'elements': ((value['elements'] as Array<any>).map(UserPublicInfoToJSON)),
        'next_start': value['next_start'],
    };
}